+++ src/main/java/org/example/backendchat/application/service/ChatService.java
@@ -1,4 +1,4 @@
-package org.example.backendchat.chat;
+package org.example.backendchat.application.service;
 
 import static org.example.backendchat.common.error.ErrorCode.*;
 
@@ -6,20 +6,19 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.example.backendchat.chat.dto.etc.ChatMessageDTO;
-import org.example.backendchat.chat.dto.req.GetAllChatMessageReqDTO;
-import org.example.backendchat.chat.dto.res.GetAllChatMessagesResDTO;
-import org.example.backendchat.chat.dto.res.GetChatMessageResDTO;
+import org.example.backendchat.application.entity.ChatMessage;
+import org.example.backendchat.application.entity.ChatRoomType;
+import org.example.backendchat.application.port.in.ChatInPort;
+import org.example.backendchat.application.port.out.ChatRepositoryOutPort;
+import org.example.backendchat.application.port.out.RedisOutPort;
+import org.example.backendchat.application.port.out.UserRepositoryOutPort;
+import org.example.backendchat.common.dto.etc.ChatMessageDTO;
+import org.example.backendchat.common.dto.req.GetAllChatMessageReqDTO;
+import org.example.backendchat.common.dto.res.GetAllChatMessagesResDTO;
+import org.example.backendchat.common.dto.res.GetChatMessageResDTO;
 import org.example.backendchat.common.exception.UserNotFoundException;
-import org.example.backendchat.entity.ChatMessage;
-import org.example.backendchat.entity.ChatRoomType;
-import org.example.backendchat.redis.RedisSubscriber;
-import org.example.backendchat.user.UserRepository;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
@@ -29,13 +28,11 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class ChatService {
+public class ChatService implements ChatInPort {
 
-	private final ChatMongoRepository chatMongoRepository;
-	private final UserRepository userRepository;
-	private final RedisTemplate<String, Object> redisTemplate;
-	private final RedisSubscriber redisSubscriber;
-	private final RedisMessageListenerContainer redisMessageListenerContainer;
+	private final RedisOutPort redisOutPort;
+	private final ChatRepositoryOutPort chatRepositoryOutPort;
+	private final UserRepositoryOutPort userRepositoryOutPort;
 
 	// 메시지 전송
 	public void sendMessage(ChatMessageDTO chatMessageDTO) {
@@ -46,29 +43,26 @@ public class ChatService {
 			.chatRoomId(chatMessageDTO.getChatRoomId())
 			.sendingTime(LocalDateTime.now().toString())
 			.build();
-		ChannelTopic channelTopic = new ChannelTopic(
-			chatMessageDTO.getChatRoomType().getType() + "/" + chatMessageDTO.getChatRoomId().toString());
-		redisMessageListenerContainer.addMessageListener(redisSubscriber, channelTopic);
+		redisOutPort.sendMessage(chatMessageDTO);
 		saveChatMessageWithAsync(chatMessage);
-		redisTemplate.convertAndSend(channelTopic.getTopic(), chatMessageDTO);
 	}
 
 	// 메시지 저장 비동기
 	@Async
 	public void saveChatMessageWithAsync(ChatMessage chatMessage) {
-		chatMongoRepository.save(chatMessage);
+		chatRepositoryOutPort.save(chatMessage);
 	}
 
 	// 메시지 조회
 	public GetAllChatMessagesResDTO getChatMessages(GetAllChatMessageReqDTO getAllChatMessageReqDTO,
 		Pageable pageable) {
-		Slice<ChatMessage> chatMessages = chatMongoRepository.findAllByChatRoomIdAndChatRoomType(
+		Slice<ChatMessage> chatMessages = chatRepositoryOutPort.findAllByChatRoomIdAndChatRoomType(
 			getAllChatMessageReqDTO.getChatRoomId(), ChatRoomType.valueOf(getAllChatMessageReqDTO.getChatRoomType()),
 			pageable);
 		List<GetChatMessageResDTO> chatMessageResDTOS = chatMessages.stream()
 			.map(chat -> GetChatMessageResDTO.builder()
 				.userId(chat.getSenderId())
-				.userName(userRepository.findById(chat.getSenderId())
+				.userName(userRepositoryOutPort.findById(chat.getSenderId())
 					.orElseThrow(() -> new UserNotFoundException("User not found", USER_NOT_FOUND))
 					.getNickName())
 				.message(chat.getMessage())
