+++ src/main/java/org/example/backendchat/common/dto/res/GetChatMessageResDTO.java
@@ -1,4 +1,4 @@
-package org.example.backendchat.chat.dto.res;
+package org.example.backendchat.common.dto.res;
 
 import lombok.Builder;
 import lombok.Getter;
